fix checkpoint

+++ train.py
@@ -1,3 +1,4 @@
+import os
 import time
 
 import torch
@@ -53,13 +54,27 @@ if __name__ == '__main__':
     optimizer = optim.Adam([{'params': net.parameters()},
                             {'params': arcFace.parameters()}],
                            lr=learning_rate, weight_decay=weight_decay)
+    scheduler = optim.lr_scheduler.MultiStepLR(optimizer, milestones=[200, 200000], gamma=0.1, last_epoch=-1)
     print(net.parameters())
     print(arcFace.parameters())
-    torch.save({'net': net.state_dict(),
-                'epoch': 0,
-                'iter': 0
-                }, modelSavePath+str(0)+'.tar')
-    print('Model saved to %s' % (modelSavePath + str(0) + '.tar'))
+    if os.path.exists(modelSavePath+'.tar'):
+        checkpoint = torch.load(modelSavePath+'.tar')
+        net.load_state_dict(checkpoint['net'])
+        arcFace.load_state_dict(checkpoint['arc'])
+        optimizer.load_state_dict(checkpoint['optimizer'])
+        scheduler.load_state_dict(checkpoint['scheduler'])
+        epoch_start = checkpoint['epoch']
+        iter_start = checkpoint['iter']
+        print('Load checkpoint Successfully!')
+        print('epoch: %d\niter: %d' % (epoch_start, iter_start))
+    else:
+        epoch_start = 0
+        iter_start = 0
+        torch.save({'net': net.state_dict(),
+                    'epoch': 0,
+                    'iter': 0
+                    }, modelSavePath+str(0)+'.tar')
+        print('Model saved to %s' % (modelSavePath + '.tar'))
 
     num_params = 0
     for param in net.parameters():
@@ -69,9 +84,8 @@ if __name__ == '__main__':
     print("Parameter Number: %d M" % (num_params / 1e6))
 
     print("Training Started!")
-    iterations = 0
-    scheduler = optim.lr_scheduler.MultiStepLR(optimizer, milestones=[200, 100000], gamma=0.1, last_epoch=-1)
-    for epoch in range(Total):
+    iterations = iter_start
+    for epoch in range(epoch_start, Total):
         data_time, train_time = 0, 0
         pred, train_x, train_y, loss = None, None, None, None
 
@@ -97,37 +111,39 @@ if __name__ == '__main__':
             train_time = train_time + tt
 
             iterations += 1
-            if iterations % 200 == 0:
-                print('Output Max pred: %s' % torch.max(feat.gather(1, train_y.view(-1, 1))))
-                print('Abs Max Gradient of net_output:', get_max_gradient(grads['feat_grad'].gather(1, train_y.view(-1, 1))))
-            if iterations % 1 == 0:
-                pred = get_label(feat)
-                acc = (pred == train_y).sum().float() / train_y.size(0) * 100
-                print('epoch: %d/%d, iters: %d, lr: %.5f, '
-                      'loss: %.5f, acc: %.5f, train_time: %.5f, data_time: %.5f' %
-                      (epoch, Total, iterations, scheduler.get_lr()[0],
-                       float(loss), float(acc), tt, dt))
-                loss_bc += float(loss)
-                acc_bc += float(acc)
+            # if iterations % 200 == 0:
+            #     print('Output Max pred: %s' % torch.max(feat.gather(1, train_y.view(-1, 1))))
+            #     print('Abs Max Gradient of net_output:',
+            #           get_max_gradient(grads['feat_grad'].gather(1, train_y.view(-1, 1))))
+
+            if iterations % 5000 == 0:
+                torch.save(net.state_dict(), modelSavePath++'_'+str(iterations)+'.pt')
+                print('Model saved to '+modelSavePath++'_'+str(iterations)+'.pt')
+            # if iterations % 1 == 0:
+            pred = get_label(feat)
+            acc = (pred == train_y).sum().float() / train_y.size(0) * 100
+            print('epoch: %d/%d, iters: %d, lr: %.5f, '
+                  'loss: %.5f, acc: %.5f, train_time: %.5f, data_time: %.5f' %
+                  (epoch, Total, iterations, scheduler.get_lr()[0],
+                   float(loss), float(acc), tt, dt))
+            loss_bc += float(loss)
+            acc_bc += float(acc)
             batch_data_time = time.time()
 
         epoch += 1
-        if epoch % 1 == 0:
-            # pred = get_label(feat)
-            # acc = (pred == train_y).sum()
-            loss_bc /= slides
-            acc_bc /= slides
-            print('epoch: %d/%d, loss: %.5f, acc: %.5f, train_time: %.5f, data_time: %.5f' %
-                  (epoch, Total, loss_bc, acc_bc, train_time, data_time))
-        if epoch > 14:
-            state = {'net': net.state_dict(),
-                     'arc': arcFace.state_dict(),
-                     'optimizer': optimizer.state_dict(),
-                     'epoch': epoch,
-                     'iter': iterations,
-                     'loss': loss_bc,
-                     'acc': acc_bc}
-            torch.save(state, modelSavePath+str(epoch)+'.tar')
-            print('Model saved to %s' % (modelSavePath+str(epoch)+'.tar'))
+        loss_bc /= slides
+        acc_bc /= slides
+        print('epoch: %d/%d, loss: %.5f, acc: %.5f, train_time: %.5f, data_time: %.5f' %
+              (epoch, Total, loss_bc, acc_bc, train_time, data_time))
+        state = {'net': net.state_dict(),
+                 'arc': arcFace.state_dict(),
+                 'optimizer': optimizer.state_dict(),
+                 'scheduler': scheduler.state_dict(),
+                 'epoch': epoch,
+                 'iter': iterations,
+                 'loss': loss_bc,
+                 'acc': acc_bc}
+        torch.save(state, modelSavePath+'.tar')
+        print('Model saved to %s' % (modelSavePath+'.tar'))
 
     print('fydnb!')
